Remove most helper functions from vds.[h|cpp] (8/8)
The helper functions specific to VDS (unit validation, coordinate
transformation etc.) moved fully into the poststack implementation. The
temporary duplicate definition of functions in vds.h was removed as well.
The helper functions will be further split up and simplified in future
commits.
