+++ internal/vds/vds.cpp
@@ -32,166 +32,6 @@ void requestdata_delete(struct requestdata* buf) {
     *buf = requestdata {};
 }
 
-int axis_todim(Axis ax) {
-    switch (ax) {
-        case I:
-        case INLINE:
-            return 0;
-        case J:
-        case CROSSLINE:
-            return 1;
-        case K:
-        case DEPTH:
-        case TIME:
-        case SAMPLE:
-            return 2;
-        default: {
-            throw std::runtime_error("Unhandled axis");
-        }
-    }
-}
-
-const std::string axis_tostring(Axis ax) {
-    switch (ax) {
-        case I:         return std::string( OpenVDS::KnownAxisNames::I()         );
-        case J:         return std::string( OpenVDS::KnownAxisNames::J()         );
-        case K:         return std::string( OpenVDS::KnownAxisNames::K()         );
-        case INLINE:    return std::string( OpenVDS::KnownAxisNames::Inline()    );
-        case CROSSLINE: return std::string( OpenVDS::KnownAxisNames::Crossline() );
-        case DEPTH:     return std::string( OpenVDS::KnownAxisNames::Depth()     );
-        case TIME:      return std::string( OpenVDS::KnownAxisNames::Time()      );
-        case SAMPLE:    return std::string( OpenVDS::KnownAxisNames::Sample()    );
-        default: {
-            throw std::runtime_error("Unhandled axis");
-        }
-    }
-}
-
-
-std::string vdsformat_tostring(OpenVDS::VolumeDataFormat format) {
-    switch (format) {
-        case OpenVDS::VolumeDataFormat::Format_U8:  return "<u1";
-        case OpenVDS::VolumeDataFormat::Format_U16: return "<u2";
-        case OpenVDS::VolumeDataFormat::Format_R32: return "<f4";
-        default: {
-            throw std::runtime_error("unsupported VDS format type");
-        }
-    }
-}
-
-/*
- * Unit validation of Z-slices
- *
- * Verify that the units of the VDS' Z axis matches the requested slice axis.
- * E.g. a Time slice is only valid if the units of the Z-axis in the VDS is
- * "Seconds" or "Milliseconds"
- */
-bool unit_validation(Axis ax, const char* zunit) {
-    /* Define some convenient lookup tables for units */
-    static const std::array< const char*, 3 > depthunits = {
-        OpenVDS::KnownUnitNames::Meter(),
-        OpenVDS::KnownUnitNames::Foot(),
-        OpenVDS::KnownUnitNames::USSurveyFoot()
-    };
-
-    static const std::array< const char*, 2 > timeunits = {
-        OpenVDS::KnownUnitNames::Millisecond(),
-        OpenVDS::KnownUnitNames::Second()
-    };
-
-    static const std::array< const char*, 1 > sampleunits = {
-        OpenVDS::KnownUnitNames::Unitless(),
-    };
-
-    auto isoneof = [zunit](const char* x) {
-        return !std::strcmp(x, zunit);
-    };
-
-    switch (ax) {
-        case I:
-        case J:
-        case K:
-        case INLINE:
-        case CROSSLINE:
-            return true;
-        case DEPTH:
-            return std::any_of(depthunits.begin(), depthunits.end(), isoneof);
-        case TIME:
-            return std::any_of(timeunits.begin(), timeunits.end(), isoneof);
-        case SAMPLE:
-            return std::any_of(sampleunits.begin(), sampleunits.end(), isoneof);
-        default: {
-            throw std::runtime_error("Unhandled axis");
-        }
-    }
-};
-
-/*
- * Until we know more about how VDS' are constructed w.r.t. to axis ordering
- * we're gonna assume that all VDS' are ordered like so:
- *
- *     voxel[0] -> depth/time/sample
- *     voxel[1] -> crossline
- *     voxel[2] -> inline
- *
- * This function will return 0 if that's not the case
- */
-bool axis_order_validation(const OpenVDS::VolumeDataLayout *layout) {
-    if (std::strcmp(layout->GetDimensionName(2), OpenVDS::KnownAxisNames::Inline())) {
-        return false;
-    }
-
-    if (std::strcmp(layout->GetDimensionName(1), OpenVDS::KnownAxisNames::Crossline())) {
-        return false;
-    }
-
-    auto z = layout->GetDimensionName(0);
-
-    /* Define some convenient lookup tables for units */
-    static const std::array< const char*, 3 > depth = {
-        OpenVDS::KnownAxisNames::Depth(),
-        OpenVDS::KnownAxisNames::Time(),
-        OpenVDS::KnownAxisNames::Sample()
-    };
-
-    auto isoneof = [z](const char* x) {
-        return !std::strcmp(x, z);
-    };
-
-    return std::any_of(depth.begin(), depth.end(), isoneof);
-}
-
-void axis_validation(Axis ax, const OpenVDS::VolumeDataLayout* layout) {
-    if (not axis_order_validation(layout)) {
-        std::string msg = "Unsupported axis ordering in VDS, expected ";
-        msg += "Depth/Time/Sample, Crossline, Inline";
-        throw std::runtime_error(msg);
-    }
-
-    auto zaxis = layout->GetAxisDescriptor(0);
-    const char* zunit = zaxis.GetUnit();
-    if (not unit_validation(ax, zunit)) {
-        std::string msg = "Unable to use " + axis_tostring(ax);
-        msg += " on cube with depth units: " + std::string(zunit);
-        throw std::runtime_error(msg);
-    }
-}
-
-nlohmann::json json_axis(
-    int voxel_dim,
-    const OpenVDS::VolumeDataLayout *layout
-) {
-    nlohmann::json doc;
-    doc = {
-        { "annotation", layout->GetDimensionName(voxel_dim)       },
-        { "min",        layout->GetDimensionMin(voxel_dim)        },
-        { "max",        layout->GetDimensionMax(voxel_dim)        },
-        { "samples",    layout->GetDimensionNumSamples(voxel_dim) },
-        { "unit",       layout->GetDimensionUnit(voxel_dim)       },
-    };
-    return doc;
-}
-
 nlohmann::json convert_axis_descriptor_to_json(
     const AxisMetadata& axis_metadata
 ) {
@@ -204,19 +44,6 @@ nlohmann::json convert_axis_descriptor_to_json(
         { "unit",       axis_metadata.unit()              },
     };
     return doc;
-}
-
-OpenVDS::ScopedVDSHandle open_vds(
-    std::string url,
-    std::string credentials
-){
-    OpenVDS::Error error;
-    OpenVDS::ScopedVDSHandle handle = OpenVDS::Open(url, credentials, error);
-
-    if(error.code != 0) {
-        throw std::runtime_error("Could not open VDS: " + error.string);
-    }
-    return handle;
 }
 
 struct requestdata handle_error(
